+++ lib/src/errors/dependency_injection_error.dart
@@ -1,0 +1,5 @@
+class DependencyInjectionError extends Error {
+  DependencyInjectionError(this.message);
+
+  final String message;
+}