+++ lib/src/logger/dependon_logger.dart
@@ -1,0 +1,13 @@
+import 'dart:developer' as developer;
+
+class DependonLogger {
+  bool enabled = false;
+
+  void log(String message) {
+    if (!enabled) {
+      return;
+    }
+
+    developer.log('Dependon: $message');
+  }
+}