+++ lib/src/models/injectable.dart
@@ -1,0 +1,5 @@
+abstract class Injectable<T extends Object> {
+  Injectable(this.builder);
+
+  T Function() builder;
+}